refactor(editor): extract bulk-import parsers to a shared include (#1200 Phase 4b)
Behaviour-preserving extraction — the FIRST step of the clean v2 bulk-import
surface. Moves the bulk-import code out of the 7,800-line legacy api.php into a
new shared include so the v2 editor API (api2.php) can reuse the SAME parser
code instead of forking it (CLAUDE.md modularity rule), per the owner's
"extract & build now" decision.

Moved VERBATIM into appWeb/public_html/includes/song_importers.php:
- 39 _bulkImport_* functions: the 8-format parsers (TXT / OpenSong / OpenLyrics
  / VideoPsalm / ProPresenter 6 / FreeShow / Proclaim / EasyWorship) + PPTX, the
  per-format process* dispatchers, the universal saver _bulkImport_saveSong, the
  ZIP walker _bulkImport_processZip, the job-progress writer _bulkImport_jobMark,
  songbook upsert, dedup (#1051), etc.
- The two shared validators _ietfBcp47Validate / _sanitiseArrangement (also
  called by non-bulk api.php handlers — still available via the require).
- The 5 BULK_IMPORT_* constants.

Only THREE lazy require paths were adjusted for the new includes/ location
(title_normalize.php, songbook-palette.php, PptxImporter.php); everything else
is byte-identical. api.php now require_once'es the include above the switch and
calls the same function names — its bulk_import_* action handlers + the
EasyWorship EXPORT helpers (_ewExport_*) are unchanged. api.php 7813 → 4378
lines; no behavioural change intended.

Verified: php -l clean on both files; every moved symbol now defined exactly
once (in the include) and zero times in api.php; api.php still requires + calls
them; the only external calls from the include are the two lazy-required helpers
+ globally-loaded getDbMysqli/getCurrentUser/logActivity (_perBookBump is a
local closure). NOTE: a live legacy bulk-import (zip + each single-file format)
should be smoke-tested in-browser — a behaviour-preserving move of live code
can't be runtime-verified headless. Working-branch commit; nothing deploys yet.
Next: v2 sync import endpoints + UI (Phase 4b.2), then async zip (4b.3).
